feat: allow configuring dashboard queries from country config
